refactor: migrate TableColumnSchemaToolbar
