+++ packages/core/client/src/modules/blocks/data-blocks/table/TableColumnSchemaToolbar.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { SchemaToolbar } from '../../schema-settings';
+import { SchemaToolbar } from '../../../../schema-settings';
 
 export const TableColumnSchemaToolbar = (props) => {
   return <SchemaToolbar initializer={false} showBorder={false} showBackground {...props} />;
